Separació en diferent arxius de les diferents funcions en Python. Substitució de la imatge de l'arxiu per una imatge representativa
